backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
bcrypt==4.3.0
boto3==1.38.2
botocore==1.38.2
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
click==8.1.8
cssselect2==0.8.0
defusedxml==0.7.1
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tinycss2==1.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
